Update for rustc

+++ src/win32/init.rs
@@ -111,7 +111,7 @@ fn init(title: Vec<u16>, builder: BuilderAttribs<'static>, builder_sharelists: O
 
             if handle.is_null() {
                 return Err(OsError(format!("CreateWindowEx function failed: {}",
-                    os::error_string(os::errno() as usize))));
+                    os::error_string(os::errno()))));
             }
 
             handle
@@ -122,7 +122,7 @@ fn init(title: Vec<u16>, builder: BuilderAttribs<'static>, builder_sharelists: O
             let hdc = unsafe { user32::GetDC(dummy_window) };
             if hdc.is_null() {
                 let err = Err(OsError(format!("GetDC function failed: {}",
-                    os::error_string(os::errno() as usize))));
+                    os::error_string(os::errno()))));
                 unsafe { user32::DestroyWindow(dummy_window); }
                 return err;
             }
@@ -139,7 +139,7 @@ fn init(title: Vec<u16>, builder: BuilderAttribs<'static>, builder_sharelists: O
                 mem::size_of::<winapi::PIXELFORMATDESCRIPTOR>() as winapi::UINT, &mut output) } == 0
             {
                 let err = Err(OsError(format!("DescribePixelFormat function failed: {}",
-                    os::error_string(os::errno() as usize))));
+                    os::error_string(os::errno()))));
                 unsafe { user32::DestroyWindow(dummy_window); }
                 return err;
             }
@@ -151,7 +151,7 @@ fn init(title: Vec<u16>, builder: BuilderAttribs<'static>, builder_sharelists: O
         unsafe {
             if gdi32::SetPixelFormat(dummy_hdc, 1, &pixel_format) == 0 {
                 let err = Err(OsError(format!("SetPixelFormat function failed: {}",
-                    os::error_string(os::errno() as usize))));
+                    os::error_string(os::errno()))));
                 user32::DestroyWindow(dummy_window);
                 return err;
             }
@@ -211,7 +211,7 @@ fn init(title: Vec<u16>, builder: BuilderAttribs<'static>, builder_sharelists: O
 
         if handle.is_null() {
             return Err(OsError(format!("CreateWindowEx function failed: {}",
-                os::error_string(os::errno() as usize))));
+                os::error_string(os::errno()))));
         }
 
         handle
@@ -222,7 +222,7 @@ fn init(title: Vec<u16>, builder: BuilderAttribs<'static>, builder_sharelists: O
         let hdc = unsafe { user32::GetDC(real_window) };
         if hdc.is_null() {
             let err = Err(OsError(format!("GetDC function failed: {}",
-                os::error_string(os::errno() as usize))));
+                os::error_string(os::errno()))));
             unsafe { user32::DestroyWindow(real_window); }
             return err;
         }
@@ -233,7 +233,7 @@ fn init(title: Vec<u16>, builder: BuilderAttribs<'static>, builder_sharelists: O
     unsafe {
         if gdi32::SetPixelFormat(hdc, 1, &pixel_format) == 0 {
             let err = Err(OsError(format!("SetPixelFormat function failed: {}",
-                os::error_string(os::errno() as usize))));
+                os::error_string(os::errno()))));
             user32::DestroyWindow(real_window);
             return err;
         }
@@ -395,7 +395,7 @@ fn create_context(extra: Option<(&gl::wgl_extra::Wgl, &BuilderAttribs<'static>)>
 
     if ctxt.is_null() {
         return Err(OsError(format!("OpenGL context creation failed: {}",
-                           os::error_string(os::errno() as usize))));
+                           os::error_string(os::errno()))));
     }
 
     Ok(ctxt as winapi::HGLRC)
@@ -508,7 +508,7 @@ fn load_opengl32_dll() -> Result<winapi::HMODULE, CreationError> {
 
     if lib.is_null() {
         return Err(OsError(format!("LoadLibrary function failed: {}",
-                                    os::error_string(os::errno() as usize))));
+                                    os::error_string(os::errno()))));
     }
 
     Ok(lib)
